+++ src/cmd/enhancement.h
@@ -29,7 +29,7 @@ public:
       string fn (filename);
       string fac(_Universe->GetFaction(faction));
       Kill();
-      _Universe->AccessCockpit()->savegame->AddUnitToSave(fn.c_str(),ENHANCEMENTPTR,fac.c_str(),(int)this);
+      _Universe->AccessCockpit()->savegame->AddUnitToSave(fn.c_str(),ENHANCEMENTPTR,fac.c_str(),(long)this);
     }
   }
